requirements.txt
numpy>=1.19.2
pandas>=1.2.0
scikit-learn>=0.24.2
joblib>=1.0.1
flask>=2.0.1
mysql-connector-python>=8.0.26
sentence-transformers>=2.2.2
torch>=2.0.0 